Removed many warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@10998 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: d4eb66fb5d2ac46c209e55b4cd829a45b880d4a2

+++ spatial/spatial.grid/src/de/uniol/inf/is/odysseus/spatial/grid/functions/IsGridFree.java
@@ -96,8 +96,8 @@ public class IsGridFree extends AbstractFunction<Boolean> {
 		int roiWidth = (int) (width / grid.cellsize);
 		int roiDepth = (int) (depth / grid.cellsize);
 
-		final int roiX = (int) Math.max(globalGridCenterX - (roiWidth / 2), 0);
-		final int roiY = (int) Math.max(globalGridCenterY - (roiDepth / 2), 0);
+		final int roiX = Math.max(globalGridCenterX - (roiWidth / 2), 0);
+		final int roiY = Math.max(globalGridCenterY - (roiDepth / 2), 0);
 
 		roiWidth = (roiX + roiWidth) > grid.width ? grid.width - roiX
 				: roiWidth;
